brain/requirements.txt
# ── gRPC Communication (The Bridge) ──
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.2

# ── LangChain AI Pipeline (The Engine) ──
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-openai==0.1.23

# ── LLM Provider ──
openai==1.51.0

# ── Utilities ──
python-dotenv==1.0.1

# ── Network Fix (Prevents httpx 0.28.0 break) ──
httpx==0.27.2